v1 lastest

+++ client.py
@@ -5,18 +5,20 @@ Created on 2017年4月14日
 @author: chenyitao
 '''
 
-import gevent.monkey
+import gevent.monkey 
 gevent.monkey.patch_all()
 import os
-os.remove('./worker.log')
+if os.path.exists('./worker.log'):
+    os.remove('./worker.log')
 
-from tddc.base import WorkerManager
+from tddc.base import WorkerManager, PackagesManager, EventCenter
 from tddc.common import TDDCLogging
 
 from parser_site import ParserSite
 from worker.parser import Parser
 from worker.storager import ParseStorager
 from worker.task import ParseTaskManager
+from worker.common.event import TDDCEventType
 
 
 class ParserManager(WorkerManager):
@@ -24,6 +26,8 @@ class ParserManager(WorkerManager):
     def __init__(self):
         TDDCLogging.info('->Parser Is Starting')
         super(ParserManager, self).__init__(ParserSite)
+        EventCenter().register(TDDCEventType.Parser.MODULES_UPDATE,
+                               PackagesManager()._models_update_event)
         self._storager = ParseStorager(ParserSite.random_hbase_node())
         self._parser = Parser()
         self._task_manager = ParseTaskManager()
